+++ custom_components/yamaha_ynca/const.py
@@ -16,8 +16,6 @@ CONF_PORT = "port"
 DATA_MODELNAME = "modelname"
 DATA_ZONES = "zones"
 
-SERVICE_SEND_RAW_YNCA = "send_raw_ynca"
-
 ATTR_COMMANDS = "commands"
 
 MANUFACTURER_NAME = "Yamaha"
